requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
autopep8==1.5.2
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.7
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-filter==2.2.0
django-js-asset==1.2.2
django-widget-tweaks==1.4.1
djangorestframework==3.11.0
gunicorn==20.0.4
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
Markdown==3.2.1
msgpack==0.6.2
multitasking==0.0.9
mysqlclient==1.4.6
numpy==1.18.4
pandas==1.0.3
Pillow==7.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-memcached==1.59
pytz==2019.3
requests==2.23.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
stripe==2.50.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
whitenoise==5.0.1
yfinance==0.1.54
zope.interface==5.1.0
bs4==0.0.1
django-allauth==0.44.0
django-rest-auth==0.9.5